+++ common/src/main/java/com/daqem/jobsplus/networking/s2c/ClientboundOpenJobsScreenPacket.java
@@ -1,18 +1,20 @@
 package com.daqem.jobsplus.networking.s2c;
 
+import java.util.List;
+
+import org.jetbrains.annotations.NotNull;
+
 import com.daqem.jobsplus.networking.JobsPlusNetworking;
 import com.daqem.jobsplus.player.job.Job;
+
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
 
 public class ClientboundOpenJobsScreenPacket implements CustomPacketPayload {
 
     private final List<Job> jobs;
-    private final int coins;
+    private final double coins;
 
     public static final StreamCodec<RegistryFriendlyByteBuf, ClientboundOpenJobsScreenPacket> STREAM_CODEC = new StreamCodec<>() {
         @Override
@@ -23,18 +25,18 @@ public class ClientboundOpenJobsScreenPacket implements CustomPacketPayload {
         @Override
         public void encode(RegistryFriendlyByteBuf buf, ClientboundOpenJobsScreenPacket packet) {
             buf.writeCollection(packet.jobs, Job.Serializer::toNetwork);
-            buf.writeInt(packet.coins);
+            buf.writeDouble(packet.coins);
         }
     };
 
-    public ClientboundOpenJobsScreenPacket(List<Job> jobs, int coins) {
+    public ClientboundOpenJobsScreenPacket(List<Job> jobs, double coins) {
         this.jobs = jobs;
         this.coins = coins;
     }
 
     public ClientboundOpenJobsScreenPacket(RegistryFriendlyByteBuf friendlyByteBuf) {
         this.jobs = friendlyByteBuf.readList(friendlyByteBuf1 -> Job.Serializer.fromNetwork(friendlyByteBuf1, null));
-        this.coins = friendlyByteBuf.readInt();
+        this.coins = friendlyByteBuf.readDouble();
     }
 
     @Override
@@ -46,7 +48,7 @@ public class ClientboundOpenJobsScreenPacket implements CustomPacketPayload {
         return jobs;
     }
 
-    public int getCoins() {
+    public double getCoins() {
         return coins;
     }
 }
