+++ common/src/main/java/com/daqem/jobsplus/networking/s2c/ClientboundOpenPowerupsScreenPacket.java
@@ -1,19 +1,21 @@
 package com.daqem.jobsplus.networking.s2c;
 
+import java.util.List;
+
+import org.jetbrains.annotations.NotNull;
+
 import com.daqem.jobsplus.networking.JobsPlusNetworking;
 import com.daqem.jobsplus.player.job.Job;
+
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.ResourceLocation;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
 
 public class ClientboundOpenPowerupsScreenPacket implements CustomPacketPayload {
 
     private final List<Job> jobs;
-    private final int coins;
+    private final double coins;
     private final ResourceLocation jobLocation;
 
     public static final StreamCodec<RegistryFriendlyByteBuf, ClientboundOpenPowerupsScreenPacket> STREAM_CODEC = new StreamCodec<>() {
@@ -25,12 +27,12 @@ public class ClientboundOpenPowerupsScreenPacket implements CustomPacketPayload 
         @Override
         public void encode(RegistryFriendlyByteBuf buf, ClientboundOpenPowerupsScreenPacket packet) {
             buf.writeCollection(packet.jobs, Job.Serializer::toNetwork);
-            buf.writeInt(packet.coins);
+            buf.writeDouble(packet.coins);
             buf.writeResourceLocation(packet.jobLocation);
         }
     };
 
-    public ClientboundOpenPowerupsScreenPacket(List<Job> jobs, int coins, ResourceLocation jobLocation) {
+    public ClientboundOpenPowerupsScreenPacket(List<Job> jobs, double coins, ResourceLocation jobLocation) {
         this.jobs = jobs;
         this.coins = coins;
         this.jobLocation = jobLocation;
@@ -38,7 +40,7 @@ public class ClientboundOpenPowerupsScreenPacket implements CustomPacketPayload 
 
     public ClientboundOpenPowerupsScreenPacket(RegistryFriendlyByteBuf friendlyByteBuf) {
         this.jobs = friendlyByteBuf.readList(friendlyByteBuf1 -> Job.Serializer.fromNetwork(friendlyByteBuf1, null));
-        this.coins = friendlyByteBuf.readInt();
+        this.coins = friendlyByteBuf.readDouble();
         this.jobLocation = friendlyByteBuf.readResourceLocation();
     }
 
@@ -51,7 +53,7 @@ public class ClientboundOpenPowerupsScreenPacket implements CustomPacketPayload 
         return jobs;
     }
 
-    public int getCoins() {
+    public double getCoins() {
         return coins;
     }
